WIP: Preview Screen Refactor
